requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.2
opencv-python-headless==4.8.0.74
opencv-contrib-python-headless==4.8.0.74
numpy==1.26.4
onnxruntime==1.16.1
mediapipe==0.10.11
faiss-cpu==1.7.4
python-multipart==0.0.6
Pillow==10.1.0
tqdm==4.66.1
joblib==1.3.2
scikit-learn==1.3.1
protobuf<4
